+++ src/lofreq/multtest.h
@@ -27,13 +27,13 @@ static char *mtc_type_str[] = {
 
 
 void
-bonf_corr(double data[], int size, int num_tests);
+bonf_corr(double data[], long int size, long int num_tests);
 
 void
-holm_bonf_corr(double data[], int size, double alpha, int num_tests);
+holm_bonf_corr(double data[], long int size, double alpha, long int num_tests);
 
-int
-fdr(double data[], int size, double alpha, int num_tests, int **irejected);
+long int
+fdr(double data[], long int size, double alpha, long int num_tests, long int **irejected);
 
 int
 mtc_str_to_type(char *t);
